Refactor publish workspace helpers into modules
