+++ scripts/publish_workspace.py
@@ -1,153 +1,34 @@
-"""Workspace preparation helpers shared by publish automation scripts.
+"""Facade module for publish workspace utilities.
 
-The routines in this module manipulate an exported copy of the repository so
-release workflows can operate on a clean tree. They are intentionally
-side-effect free apart from file-system writes inside the provided workspace
-root, which keeps the publish scripts deterministic and safe to run locally.
+This module preserves the historical import surface while delegating the
+implementation to focused helper modules. Callers should prefer the re-exported
+functions documented in the respective modules for clarity.
 """
 
 from __future__ import annotations
 
-import tarfile
-import tempfile
 import typing as typ
-from pathlib import Path
 
-import tomllib
-from plumbum import local
-from publish_patch import REPLACEMENTS, apply_replacements
-from tomlkit import array, dumps, parse
-from tomlkit.items import Array
+import publish_workspace_archive as _archive
+import publish_workspace_dependencies as _dependencies
+import publish_workspace_members as _members
+import publish_workspace_patch as _patch
+import publish_workspace_serialise as _serialise
+import publish_workspace_versioning as _versioning
+from plumbum import local as _default_local
+from tomlkit import parse as _default_parse
 
 if typ.TYPE_CHECKING:
-    from tomlkit.toml_document import TOMLDocument
+    from pathlib import Path
 
-PROJECT_ROOT = Path(__file__).resolve().parents[1]
-
-PUBLISHABLE_CRATES: typ.Final[tuple[str, ...]] = (
-    "rstest-bdd-patterns",
-    "rstest-bdd-macros",
-    "rstest-bdd",
-    "cargo-bdd",
-)
+local = _default_local
+parse = _default_parse
 
 
 def export_workspace(destination: Path) -> None:
-    """Extract the repository HEAD into ``destination`` via ``git archive``.
-
-    Parameters
-    ----------
-    destination : Path
-        Directory that will receive the exported workspace contents.
-
-    Returns
-    -------
-    None
-        The repository snapshot is written directly to ``destination``.
-    """
-    with tempfile.TemporaryDirectory() as archive_dir:
-        archive_path = Path(archive_dir) / "workspace.tar"
-        git_archive = local["git"][
-            "archive", "--format=tar", "HEAD", f"--output={archive_path}"
-        ]
-        with local.cwd(PROJECT_ROOT):
-            return_code, stdout, stderr = git_archive.run(
-                timeout=60,
-                retcode=None,
-            )
-        if return_code != 0:
-            diagnostics = (stderr or stdout or "").strip()
-            detail = f": {diagnostics}" if diagnostics else ""
-            message = f"git archive failed with exit code {return_code}{detail}"
-            raise SystemExit(message)
-        with tarfile.open(archive_path) as tar:
-            tar.extractall(destination, filter="data")
-
-
-def strip_patch_section(manifest: Path) -> None:
-    """Strip the ``[patch.crates-io]`` section from ``manifest``.
-
-    Parameters
-    ----------
-    manifest : Path
-        Manifest whose patch entries should be removed in place.
-
-    Returns
-    -------
-    None
-        The manifest on disk is rewritten without the patch section.
-    """
-    document = parse(manifest.read_text(encoding="utf-8"))
-    patch_tables = _get_patch_crates_io_tables(document)
-    if patch_tables is None:
-        return
-
-    patch_table, _ = patch_tables
-    _remove_patch_section(document, patch_table)
-    _write_manifest_with_newline(document, manifest)
-
-
-def prune_workspace_members(manifest: Path) -> None:
-    """Remove non-crate entries from the workspace members list.
-
-    Parameters
-    ----------
-    manifest : Path
-        Workspace manifest whose members array should only contain crates.
-
-    Returns
-    -------
-    None
-        The manifest is rewritten with only crate directories listed.
-    """
-    document = parse(manifest.read_text(encoding="utf-8"))
-    members = _get_valid_workspace_members(document)
-    if members is None:
-        return
-
-    changed = _filter_workspace_members(members)
-    _write_manifest_if_changed(
-        document=document,
-        manifest=manifest,
-        changed=changed,
-        members=members,
-    )
-
-
-def _get_valid_workspace_members(document: TOMLDocument) -> Array | None:
-    """Return the workspace members array when it exists and is valid."""
-    workspace = document.get("workspace")
-    if workspace is None:
-        return None
-
-    members = workspace.get("members")
-    if members is None:
-        return None
-
-    return _ensure_members_array(workspace, members)
-
-
-def _filter_workspace_members(members: Array) -> bool:
-    """Remove ineligible workspace members, returning ``True`` if mutated."""
-    changed = False
-    for index in range(len(members) - 1, -1, -1):
-        entry = members[index]
-        if not isinstance(entry, str) or Path(entry).name not in PUBLISHABLE_CRATES:
-            del members[index]
-            changed = True
-
-    return changed
-
-
-def _write_manifest_if_changed(
-    *, document: TOMLDocument, manifest: Path, changed: bool, members: Array
-) -> None:
-    """Persist ``document`` to ``manifest`` only when ``changed`` is ``True``."""
-    if not _should_write_manifest(changed=changed, document=document):
-        return
-
-    _format_multiline_members_if_needed(members)
-    _write_manifest_with_newline(document, manifest)
+    """Dispatch to the archive helper while respecting runtime monkeypatches."""
+    _archive.local = local
+    _archive.export_workspace(destination)
 
 
 def apply_workspace_replacements(
@@ -157,229 +38,80 @@ def apply_workspace_replacements(
     include_local_path: bool,
     crates: tuple[str, ...] | None = None,
 ) -> None:
-    """Rewrite workspace dependency declarations for publish workflows.
-
-    Parameters
-    ----------
-    workspace_root : Path
-        Root of the exported workspace containing crate directories.
-    version : str
-        Version string written to dependency entries.
-    include_local_path : bool
-        When ``True`` the relative ``path`` entries are retained so dry-run
-        checks use the local workspace.
-    crates : tuple[str, ...] | None, optional
-        Specific crates to update. Defaults to all known crates when ``None``.
-
-    Returns
-    -------
-    None
-        Each targeted manifest is rewritten in place.
-    """
-    targets = REPLACEMENTS if crates is None else crates
-    for crate in targets:
-        if crate not in REPLACEMENTS:
-            continue
-        manifest = workspace_root / "crates" / crate / "Cargo.toml"
-        apply_replacements(
-            crate,
-            manifest,
-            version,
-            include_local_path=include_local_path,
-        )
+    """Proxy dependency rewriting through the dedicated module."""
+    _dependencies.apply_workspace_replacements(
+        workspace_root,
+        version,
+        include_local_path=include_local_path,
+        crates=crates,
+    )
 
 
-def workspace_version(manifest: Path) -> str:
-    """Return the workspace package version from the root manifest.
-
-    Parameters
-    ----------
-    manifest : Path
-        Path to the workspace ``Cargo.toml`` file.
-
-    Returns
-    -------
-    str
-        The semantic version configured under ``[workspace.package]``.
-
-    Raises
-    ------
-    SystemExit
-        Raised when the workspace manifest lacks the version entry.
-    """
-    manifest_text = manifest.read_text(encoding="utf-8")
-    data = tomllib.loads(manifest_text)
-    try:
-        return data["workspace"]["package"]["version"]
-    except KeyError as err:
-        message = (
-            f"expected [workspace.package].version in {manifest}; "
-            "[workspace.package] must define a version for publish automation to run. "
-            "Check the manifest defines the key."
-        )
-        if snippet := _workspace_section_excerpt(manifest_text):
-            indented_snippet = "\n".join(f"    {line}" for line in snippet)
-            message = f"{message}\n\nWorkspace manifest excerpt:\n{indented_snippet}"
-        raise SystemExit(message) from err
+def prune_workspace_members(manifest: Path) -> None:
+    """Filter workspace members using the current ``parse`` implementation."""
+    _members.parse = parse
+    _members.prune_workspace_members(manifest)
 
 
-def _workspace_section_excerpt(manifest_text: str) -> list[str] | None:
-    """Return the lines around the ``[workspace]`` section for diagnostics."""
-    lines = manifest_text.splitlines()
-    workspace_index = _find_workspace_section_index(lines)
-
-    if workspace_index is None:
-        return None
-
-    return _extract_section_lines(lines, workspace_index)
-
-
-def _find_workspace_section_index(lines: list[str]) -> int | None:
-    """Find the index of the [workspace] section."""
-    for index, line in enumerate(lines):
-        if line.strip().startswith("[workspace"):
-            return index
-    return None
-
-
-def _extract_section_lines(lines: list[str], workspace_index: int) -> list[str]:
-    """Extract lines around the workspace section for diagnostics."""
-    start = max(workspace_index - 1, 0)
-    end = workspace_index + 1
-
-    while _should_include_more_lines(lines, end, start):
-        end += 1
-
-    return lines[start:end]
+def strip_patch_section(manifest: Path) -> None:
+    """Remove the patch section using the current ``parse`` implementation."""
+    _patch.parse = parse
+    _patch.strip_patch_section(manifest)
 
 
 def remove_patch_entry(manifest: Path, crate: str) -> None:
-    """Remove the ``crate`` entry from the root ``[patch.crates-io]`` table.
-
-    Parameters
-    ----------
-    manifest : Path
-        Root workspace manifest that may contain patch overrides.
-    crate : str
-        Name of the crate whose patch entry should be removed.
-
-    Returns
-    -------
-    None
-        The manifest is rewritten only when the patch entry was present.
-    """
-    document = parse(manifest.read_text(encoding="utf-8"))
-    patch_tables = _get_patch_crates_io_tables(document)
-    if patch_tables is None:
-        return
-
-    patch_table, crates_io = patch_tables
-    if crate not in crates_io:
-        return
-
-    _remove_crate_and_cleanup_empty_sections(
-        document=document,
-        patch_table=patch_table,
-        crates_io=crates_io,
-        crate=crate,
-    )
-    _write_manifest_with_newline(document, manifest)
+    """Remove a single patch entry using the current ``parse`` implementation."""
+    _patch.parse = parse
+    _patch.remove_patch_entry(manifest, crate)
 
 
-def _get_patch_crates_io_tables(
-    document: TOMLDocument,
-) -> tuple[dict[str, typ.Any], dict[str, typ.Any]] | None:
-    """Return the patch and crates-io tables when both are present."""
-    patch_table = document.get("patch")
-    if patch_table is None:
-        return None
-
-    crates_io = patch_table.get("crates-io")
-    if crates_io is None:
-        return None
-
-    return (
-        typ.cast("dict[str, typ.Any]", patch_table),
-        typ.cast("dict[str, typ.Any]", crates_io),
-    )
+def workspace_version(manifest: Path) -> str:
+    """Expose the workspace version helper."""
+    return _versioning.workspace_version(manifest)
 
 
-def _remove_patch_section(
-    document: TOMLDocument, patch_table: dict[str, typ.Any]
-) -> None:
-    """Remove the entire ``[patch.crates-io]`` table from ``document``."""
-    patch_table.pop("crates-io", None)
-    if not patch_table:
-        del document["patch"]
+PUBLISHABLE_CRATES: typ.Final[tuple[str, ...]] = _members.PUBLISHABLE_CRATES
+_get_valid_workspace_members = _members._get_valid_workspace_members
+_ensure_members_array = _members._ensure_members_array
+_convert_list_to_array = _members._convert_list_to_array
+_filter_workspace_members = _members._filter_workspace_members
+_should_write_manifest = _members._should_write_manifest
+_format_multiline_members_if_needed = _members._format_multiline_members_if_needed
+_write_manifest_if_changed = _members._write_manifest_if_changed
+_get_patch_crates_io_tables = _patch._get_patch_crates_io_tables
+_remove_patch_section = _patch._remove_patch_section
+_remove_crate_and_cleanup_empty_sections = (
+    _patch._remove_crate_and_cleanup_empty_sections
+)
+_write_manifest_with_newline = _serialise._write_manifest_with_newline
+_workspace_section_excerpt = _versioning._workspace_section_excerpt
+_find_workspace_section_index = _versioning._find_workspace_section_index
+_extract_section_lines = _versioning._extract_section_lines
+_should_include_more_lines = _versioning._should_include_more_lines
 
-
-def _remove_crate_and_cleanup_empty_sections(
-    *,
-    document: TOMLDocument,
-    patch_table: dict[str, typ.Any],
-    crates_io: dict[str, typ.Any],
-    crate: str,
-) -> None:
-    """Remove ``crate`` from the patch section and drop empty tables."""
-    del crates_io[crate]
-    if not crates_io:
-        del patch_table["crates-io"]
-    if not patch_table:
-        del document["patch"]
-
-
-def _should_include_more_lines(lines: list[str], end: int, start: int) -> bool:
-    """Return ``True`` when diagnostic extraction should continue."""
-    if end >= len(lines):
-        return False
-
-    if end - start >= 8:
-        return False
-
-    stripped = lines[end].strip()
-    is_non_workspace_section = stripped.startswith("[") and not stripped.startswith(
-        "[workspace"
-    )
-    return not is_non_workspace_section
-
-
-def _ensure_members_array(workspace: dict, members: object) -> Array | None:
-    """Normalise ``workspace['members']`` to a TOML array when possible."""
-    if isinstance(members, Array):
-        return members
-
-    if isinstance(members, list):
-        return _convert_list_to_array(workspace, members)
-
-    return None
-
-
-def _convert_list_to_array(workspace: dict, members: list) -> Array:
-    """Convert ``members`` list to a TOML array attached to ``workspace``."""
-    rebuilt_members = array()
-    rebuilt_members.extend(members)
-    workspace["members"] = rebuilt_members
-    return typ.cast("Array", rebuilt_members)
-
-
-def _should_write_manifest(*, changed: bool, document: TOMLDocument) -> bool:
-    """Return ``True`` when the manifest should be persisted."""
-    if not changed:
-        return False
-
-    return document.get("workspace") is not None
-
-
-def _format_multiline_members_if_needed(members: Array) -> None:
-    """Ensure ``members`` is rendered as multiline when it spans lines."""
-    if "\n" in members.as_string():
-        members.multiline(multiline=True)
-
-
-def _write_manifest_with_newline(document: TOMLDocument, manifest: Path) -> None:
-    """Serialise ``document`` to ``manifest`` and ensure a trailing newline."""
-    rendered = dumps(document)
-    if not rendered.endswith("\n"):
-        rendered = f"{rendered}\n"
-
-    manifest.write_text(rendered, encoding="utf-8")
+__all__ = [
+    "PUBLISHABLE_CRATES",
+    "_convert_list_to_array",
+    "_ensure_members_array",
+    "_extract_section_lines",
+    "_filter_workspace_members",
+    "_find_workspace_section_index",
+    "_format_multiline_members_if_needed",
+    "_get_patch_crates_io_tables",
+    "_get_valid_workspace_members",
+    "_remove_crate_and_cleanup_empty_sections",
+    "_remove_patch_section",
+    "_should_include_more_lines",
+    "_should_write_manifest",
+    "_workspace_section_excerpt",
+    "_write_manifest_if_changed",
+    "_write_manifest_with_newline",
+    "apply_workspace_replacements",
+    "export_workspace",
+    "local",
+    "parse",
+    "prune_workspace_members",
+    "remove_patch_entry",
+    "strip_patch_section",
+    "workspace_version",
+]
